Create Note.py
Note class constructs objects that have a letter note name (eg. C#7) associated with a number (currently within the MIDI tuning range: 21 (A0) to 128 (G#9).

I am not sure if it is more appropriate for this to be a data structure than a class... if that's a thing in python. Should investigate.